Trait oldnav_lib::navdata::geohash::Rectifiable

pub trait Rectifiable {
    fn spherical_rectify(&mut self, range: &Bounds);
}

A trait for a vector type object that can be rectified to fit within a boundary.

Required Methods

fn spherical_rectify(&mut self, range: &Bounds)

wrap this position around the bounds, such that it becomes the equivalent position within the bounds.